- Update deps. - Begin refactoring into a less-monolithic source code organization on the way to Full Typescript. - Seal typescript@4.9.5 in package.json due to a problem compiling src/external/brython.js via Typescript (this will be fixed later).
