+++ VERSIONS.md
@@ -66,3 +66,4 @@
 - **1.0.63** - Update deps.
 - **1.0.64** - Update deps. Improve syntax highlighting for inline code rendering. Fix issue with 'bash' code blocks not being highlighted without a language identifier. 
 - **1.0.65** - Update deps. Add option to 'configure()' to allow overriding of the default 'xypicURL', to support at least the 'smartdown/editor' project.
+- **1.0.66** - Update deps. Begin refactoring into a less-monolithic source code organization on the way to Full Typescript. Seal typescript@4.9.5 in package.json due to a problem compiling src/external/brython.js via Typescript (this will be fixed later).
