+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "smartdown",
-  "version": "1.0.65",
+  "version": "1.0.66",
   "description": "A library for translating, rendering and interacting with Smartdown documents. Smartdown is an extension of Markdown that provides richer media support and reactive programming capability.",
   "main": "dist/lib/smartdown.js",
   "scripts": {
@@ -47,33 +47,33 @@
   },
   "homepage": "https://github.com/smartdown/smartdown#readme",
   "devDependencies": {
-    "@babel/core": "^7.22.9",
+    "@babel/core": "^7.22.11",
     "@babel/plugin-syntax-dynamic-import": "^7.8.3",
     "@babel/plugin-transform-modules-umd": "7.22.5",
-    "@babel/plugin-transform-runtime": "^7.22.9",
-    "@babel/preset-env": "^7.22.9",
+    "@babel/plugin-transform-runtime": "^7.22.10",
+    "@babel/preset-env": "^7.22.10",
     "@typescript-eslint/eslint-plugin": "^5.60.1",
     "@typescript-eslint/parser": "^5.60.1",
-    "autoprefixer": "^10.4.14",
+    "autoprefixer": "^10.4.15",
     "babel-eslint": "^10.1.0",
     "babel-loader": "^9.1.3",
-    "chromedriver": "^114.0.2",
+    "chromedriver": "^116.0.0",
     "copy-webpack-plugin": "^11.0.0",
     "css-loader": "6.8.1",
     "docma": "^3.2.2",
-    "eslint": "^8.45.0",
+    "eslint": "^8.48.0",
     "eslint-config-airbnb-typescript": "^17.1.0",
-    "eslint-plugin-import": "^2.27.5",
+    "eslint-plugin-import": "^2.28.1",
     "file-loader": "^6.2.0",
     "filemanager-webpack-plugin": "^8.0.0",
-    "fsevents": "^2.3.2",
+    "fsevents": "^2.3.3",
     "html-webpack-plugin": "^5.5.3",
     "http-server": "^14.1.1",
-    "instant-mocha": "^1.5.1",
+    "instant-mocha": "^1.5.2",
     "jsdom": "^22.1.0",
     "mini-css-extract-plugin": "^2.7.6",
     "mkcert": "^1.5.1",
-    "nightwatch": "^3.0.1",
+    "nightwatch": "^3.1.3",
     "node-sass": "^9.0.0",
     "null-loader": "^4.0.1",
     "postcss-loader": "7.3.3",
@@ -83,7 +83,7 @@
     "source-map": "^0.7.4",
     "style-loader": "^3.3.3",
     "ts-loader": "^9.4.4",
-    "typescript": "^5.1.6",
+    "typescript": "4.9.5",
     "webpack": "^5.88.2",
     "webpack-bundle-analyzer": "^4.9.0",
     "webpack-cli": "^5.1.4",
@@ -91,12 +91,12 @@
     "webpack-node-externals": "^3.0.0"
   },
   "dependencies": {
-    "@babel/runtime": "^7.22.6",
+    "@babel/runtime": "^7.22.11",
     "@stdlib/dist-datasets-sotu": "0.0.96",
     "@stdlib/dist-datasets-tree-exclude": "0.0.96",
     "@stdlib/dist-tree": "0.0.96",
     "@webcomponents/webcomponentsjs": "^2.8.0",
-    "axios": "^1.4.0",
+    "axios": "^1.5.0",
     "base64-js": "^1.5.1",
     "brython": "^3.11.3",
     "dompurify": "^3.0.5",
