+++ src/extensions/Mermaid.js
@@ -9,7 +9,7 @@ Mermaid.register = function register() {
   registerExtension(
     'mermaid',
     [
-      'https://unpkg.com/mermaid/dist/mermaid.min.js',
+      'https://unpkg.com/mermaid@9.4/dist/mermaid.min.js',
       function() {
         console.log('mermaid loaded');
         initializeMermaid();
